Conflicts:
	core/.classpath

+++ core/src/main/java/com/contento3/security/group/service/impl/GroupServiceImpl.java
@@ -1,18 +1,14 @@
 package com.contento3.security.group.service.impl;
 
 import java.util.Collection;
-
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import com.contento3.cms.page.dto.PageDto;
-import com.contento3.cms.page.template.dto.TemplateDto;
 import com.contento3.security.group.dao.GroupDao;
 import com.contento3.security.group.dto.GroupDto;
-import com.contento3.security.group.model.Group;
 import com.contento3.security.group.service.GroupAssembler;
 import com.contento3.security.group.service.GroupService;
 import com.contento3.security.group.service.impl.GroupServiceImpl;
+import com.contento3.security.user.dto.SaltedHibernateUserDto;
 
 
 @Transactional(readOnly = true, propagation = Propagation.REQUIRES_NEW)
@@ -61,13 +57,23 @@ public class GroupServiceImpl implements GroupService {
 		groupDao.update(groupAssembler.dtoToDomain(groupDto));
 		
 	}
+	
 
 	@Transactional(readOnly = false, propagation = Propagation.REQUIRES_NEW)
 	@Override
-	public void delete(GroupDto group) {
-		groupDao.delete(groupAssembler.dtoToDomain(group));
+	public void delete(GroupDto group) throws Exception {
 		
+		
+		Collection<SaltedHibernateUserDto> members;
+		members=group.getMembers();
+		
+		if(members.isEmpty())
+			groupDao.delete(groupAssembler.dtoToDomain(group));
+		else
+			throw new Exception("can`t delete user associated to group");		
 	}
+
+	
 	
 	
 }
